tests/wpt/web-platform-tests/workers/interfaces/SharedWorkerGlobalScope/onconnect.html

<!doctype html>
<title>onconnect</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id="log"></div>
<script>
async_test(function() {
  var w1 = new SharedWorker('onconnect.js', '');
  w1.port.addEventListener('message', this.step_func(function(e) {
    assert_array_equals(e.data, ['null', 'null', 'function', '']);
  }), false);
  w1.port.start();
});
</script>
